# Remove the use of fmt.Sprintf